Add data systems in Project context
