+++ Assets/_Project/Prefabs/Installers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb74bf26af69bf84ea8181c030f4e247
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
