+++ Assets/_Project/Prefabs/Installers/GameDataInstaller.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 17925df43211e5f40b61f3af2a810b29
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
