+++ Assets/_Project/Scripts/Interfaces/Data/IResourceStorageData.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using Project.Systems.Data;
+
+namespace Project.Interfaces.Data
+{
+    public interface IResourceStorageData : ISaveable
+    {
+        List<GameResourceData> Resources { get; }
+    }
+}
