+++ Assets/_Project/Scripts/Interfaces/Data/IResourceStorageData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45788e8ef57cd9649a968ccbdcb15e09
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
