+++ Assets/_Project/Scripts/Interfaces/Data/ISaveable.cs
@@ -1,0 +1,7 @@
+﻿namespace Project.Interfaces.Data
+{
+    public interface ISaveable
+    {
+        void Save();
+    }
+}
