+++ Assets/_Project/Scripts/Interfaces/Data/ISaveable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 05be5429a191612499d524bb07c894d4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
