+++ Assets/_Project/Scripts/Systems/Data/GameData.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Project.Systems.Data
+{
+    [System.Serializable]
+    public class GameData
+    {
+        public List<GameResourceData> Storage = new List<GameResourceData>();
+    }
+}
