+++ Assets/_Project/Scripts/Systems/Data/GameData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec4715efe9d3e074cb4f75c159402bce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
