+++ Assets/_Project/Scripts/Systems/Data/GameDataService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92d159b5e04946e479c28cd11c90e630
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
