+++ Assets/_Project/Scripts/Systems/Data/GameResourceData.cs
@@ -1,0 +1,9 @@
+﻿namespace Project.Systems.Data
+{
+    [System.Serializable]
+    public class GameResourceData
+    {
+        public int ID;
+        public int Value;
+    }
+}
