+++ Assets/_Project/Scripts/Systems/Data/GameResourceData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0de93b850c3511b43bbaefeedac411ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
