+++ Assets/_Project/Scripts/Systems/Data/ResourceStorageProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dfc17899c61759842bad8ef7be2be2c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
