+++ Assets/_Project/Scripts/Utils/PlayerPrefsReseter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09e16b95474c5c443a4260c086ae732d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
