+++ Assets/_Project/Scripts/_TEST.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f43a9a459de04b845ab870df80b56db4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
